Initial account handling. #22
