+++ package.json
@@ -7,12 +7,14 @@
   },
   "dependencies": {
     "axios": "^1.7.9",
+    "bcryptjs": "^2.4.3",
     "dotenv": "^16.4.7",
     "ejs": "^3.1.10",
     "express": "^4.21.2",
     "express-rate-limit": "^7.4.1",
     "fs": "^0.0.1-security",
     "https-proxy-agent": "^7.0.6",
+    "jsonwebtoken": "^9.0.2",
     "mongodb": "^6.11.0",
     "path": "^0.12.7",
     "vhost": "^3.0.2"
