+++ views/partials/forms.ejs
@@ -1,0 +1,136 @@
+  <header>
+    <form action="/" method="get" id="filter">
+
+      <input type="hidden" name="page" value="<%= page %>">
+      <input type="hidden" name="ascending" value="<%= ascending %>">
+
+      <div class="filterSelection">
+        <div class="mainSelection">
+          <button type="submit" style="display: none"></button>
+
+          <select name="category" id="category" onchange="applyFilters(true)">
+            <option value="null" <%= category === 'null' ? 'selected' : '' %>>Velg kategori</option>
+            <option value="aromatisert" <%= category === 'aromatisert' ? 'selected' : '' %>>Aromatisert vin</option>
+            <option value="brennevin" <%= category === 'brennevin' ? 'selected' : '' %>>Brennevin</option>
+            <option value="fruktvin" <%= category === 'fruktvin' ? 'selected' : '' %>>Fruktvin</option>
+            <option value="hvitvin" <%= category === 'hvitvin' ? 'selected' : '' %>>Hvitvin</option>
+            <option value="mjød" <%= category === 'mjød' ? 'selected' : '' %>>Mjød</option>
+            <option value="musserende" <%= category === 'musserende' ? 'selected' : '' %>>Musserende vin</option>
+            <option value="perlende" <%= category === 'perlende' ? 'selected' : '' %>>Perlende vin</option>
+            <option value="rosévin" <%= category === 'rosévin' ? 'selected' : '' %>>Rosévin</option>
+            <option value="rødvin" <%= category === 'rødvin' ? 'selected' : '' %>>Rødvin</option>
+            <option value="sake" <%= category === 'sake' ? 'selected' : '' %>>Sake</option>
+            <option value="sider" <%= category === 'sider' ? 'selected' : '' %>>Sider</option>
+            <option value="sterkvin" <%= category === 'sterkvin' ? 'selected' : '' %>>Sterkvin</option>
+            <option value="øl" <%= category === 'øl' ? 'selected' : '' %>>Øl</option>
+          </select>
+
+          <input type="search" id="ssearch" name="storelike" value="<%= storelike || null %>" placeholder="Butikk (e.g., Stavanger) ...">
+
+          <div>
+            <select name="sort" id="sort" onchange="applyFilters(true)">
+              <option value="discount" <%= sort === 'discount' ? 'selected' : '' %>>Prisendring</option>
+              <option value="price" <%= sort === 'price' ? 'selected' : '' %>>Pris</option>
+              <option value="literprice" <%= sort === 'literprice' ? 'selected' : '' %>>Literpris</option>
+              <option value="alcoholprice" <%= sort === 'alcoholprice' ? 'selected' : '' %>>Alkoholpris</option>
+              <option value="alcohol" <%= sort === 'alcohol' ? 'selected' : '' %>>Alkoholprosent</option>
+            </select>
+
+            <button id="sortButton">
+              <%= ascending === true ? '↑' : '↓' %>
+            </button>
+          </div>
+        </div>
+
+        <div class="mainButtons">
+          <button id="info">informasjon</button>
+          <button id="clearFilters">nullstill</button>
+          <button id="toggleAdvanced">flere valg</button>
+        </div>
+      </div>
+
+      <div id="advancedSelection" class="advancedSelection" style="display: none;">
+        <input type="search" id="nsearch" name="search" value="<%= search || null %>" placeholder="Søk etter navn ...">
+
+        <select name="store" id="store" onchange="applyFilters(true)">
+          <!-- Options will be dynamically populated by `public/scripts/stores.js`. -->
+        </select>
+
+        <select name="country" id="country" onchange="applyFilters(true)">
+          <!-- Options will be dynamically populated by `public/scripts/countries.js`. -->
+        </select>
+
+        <div>
+          <label>Vol. over</label>
+          <input type="float" id="fvolume" name="volume" min="0" step="1" value="<%= volume || null %>">
+          <label>cL</label>
+        </div>
+
+        <div>
+          <label>Alk. over</label>
+          <input type="float" id="falcohol" name="alcohol" min="0" step="0.5" value="<%= alcohol || null %>">
+          <label id="percent">%</label>
+        </div>
+
+        <div>
+          <label>Årg. tom.</label>
+          <input type="number" id="iyear" name="year" min="1900" step="10" value="<%= year || null %>">
+          <label>a.D.</label>
+        </div>
+      </div>
+    </form>
+  </header>
+
+  <div id="selectCookies" class="modal">
+    <div class="selectCookies">
+      <p style="line-height: 1.5;">
+        For å gjøre <b>SNUBLEJUICE.no</b> enda bedre ønsker vi å samle inn informasjonskapsler. Ved å godta, samtykker du til bruk av informasjonskapsler.
+      </p>
+      <div class="cookieSelection">
+        <button id="decline" onclick="declineCookies()">ELLERS TAKK</button>
+        <img src="images/800x800.png" alt="SNUBLEJUICE.no">
+        <button id="accept" onclick="acceptCookies()">DEN ER GREI</button>
+      </div>
+    </div>
+  </div>
+
+  <div id="infobox" class="infobox" style="display: none;">
+    <div class="head">
+      <b>Hva er SNUBLEJUICE?</b>
+      <span class="exit">x</span>
+    </div>
+    <div style="line-height: 1.5;">
+      Her finner du Vinmonopolets hemmelige tilbud samt varers prishistorikk.
+    </div>
+    <div style="line-height: 1.5;">
+      Vinmonopolet justerer nemlig prisene hvert månedsskifte. For å gjøre livet ditt litt enklere holder vi styr på prishistorikken til varer, og gir en oversikt over tilbudsvarer.
+    </div>
+    <div style="line-height: 1.5;">
+      Lagerbeholdning oppdateres jevnlig, så avvik kan forekomme. Alltid dobbeltsjekk på produktets side.
+    </div>
+  </div>
+
+  <div class="pagination" id="top">
+    <div onclick="window.location.href='/?fresh=false'">SNUBLEJUICE.no</div>
+    <img src="images/snublejuice.png" alt="SNUBLEJUICE.no" onclick="window.location.href='/?fresh=false'">
+
+    <div class="pages">
+      <% if (page > 1) { %>
+          <a href="#" onclick="changePage(<%= page - 1 %>)"><-</a>
+      <% } %>
+
+      <span><%= page %> av <%= totalPages %></span>
+
+      <% if (page < totalPages) { %>
+          <a href="#" onclick="changePage(<%= page + 1 %>)">-></a>
+      <% } %>
+    </div>
+  </div>
+
+  <% if (new Date().getDate() <= 5) { %>
+    <div class="message">[!] NY MÅNED; NYE TILBUD [!]</div>
+  <% } %>
+  <div id="message-stores" style="display: none;">
+      <div class="message">N.b.: Faktisk lagerbeholdning kan avvike. Oppdateres med jevne mellomrom.</div>
+      <div class="message">Sist oppdatert <%= updated %>.</div>
+  </div>
