+++ views/products.ejs
@@ -24,7 +24,8 @@
 </head>
 
 <body>
-  <%- include('partials/header', { updated }) %>
+  <%- include('partials/header') %>
+  <%- include('partials/forms', { updated }) %>
 
   <% if (data.length === 0) { %>
     <p style="text-align: center;">Ingen resultater finnes for valgt kominasjon.</p>
